code refactor: declarations moved from .cpp to komodo_defs.h or else
