+++ src/cc/CCcustom.cpp
@@ -14,6 +14,7 @@
  ******************************************************************************/
 
 #include "key_io.h"
+
 #include "CCinclude.h"
 #include "CCassets.h"
 #include "CCfaucet.h"
