+++ src/crosschain.cpp
@@ -43,8 +43,6 @@
 
 
 int NOTARISATION_SCAN_LIMIT_BLOCKS = 1440;
-CBlockIndex *komodo_getblockindex(uint256 hash);
-
 
 /* On KMD */
 uint256 CalculateProofRoot(const char* symbol, uint32_t targetCCid, int kmdHeight,
