+++ src/deprecation.cpp
@@ -25,9 +25,9 @@
 #include "ui_interface.h"
 #include "util.h"
 #include "chainparams.h"
+#include "komodo_defs.h"
 
 static const std::string CLIENT_VERSION_STR = FormatVersion(CLIENT_VERSION);
-extern char ASSETCHAINS_SYMBOL[KOMODO_ASSETCHAIN_MAXLEN];
 
 void EnforceNodeDeprecation(int nHeight, bool forceLogging, bool fThread) {
