+++ src/init.cpp
@@ -23,6 +23,7 @@
 #endif
 
 #include "init.h"
+#include "alert.h"
 #include "crypto/common.h"
 #include "primitives/block.h"
 #include "addrman.h"
@@ -93,14 +94,6 @@
 using namespace std;
 
 #include "komodo_defs.h"
-extern void ThreadSendAlert();
-extern bool komodo_dailysnapshot(int32_t height);
-extern int32_t KOMODO_LOADINGBLOCKS;
-extern bool VERUS_MINTBLOCKS;
-extern char ASSETCHAINS_SYMBOL[];
-extern int32_t KOMODO_SNAPSHOT_INTERVAL;
-
-extern void komodo_init(int32_t height);
 
 ZCJoinSplit* pzcashParams = NULL;
 
@@ -753,10 +746,6 @@ void ThreadNotifyRecentlyAdded()
         mempool.NotifyRecentlyAdded();
     }
 }
-
-/* declarations needed for ThreadUpdateKomodoInternals */
-void komodo_passport_iteration();
-void komodo_cbopretupdate(int32_t forceflag);
 
 void ThreadUpdateKomodoInternals() {
     RenameThread("int-updater");
