+++ src/komodo_bitcoind.h
@@ -19,21 +19,13 @@
 #include <curl/easy.h>
 #include "primitives/nonce.h"
 #include "consensus/params.h"
-#include "komodo_defs.h"
 #include "script/standard.h"
+#include "init.h"
+#include "main.h"
+#include "komodo_defs.h"
 #include "cc/CCinclude.h"
 
 const char *LOG_KOMODOBITCOIND = "komodostaking";
-
-int32_t komodo_notaries(uint8_t pubkeys[64][33],int32_t height,uint32_t timestamp);
-int32_t komodo_electednotary(int32_t *numnotariesp,uint8_t *pubkey33,int32_t height,uint32_t timestamp);
-int32_t komodo_voutupdate(bool fJustCheck,int32_t *isratificationp,int32_t notaryid,uint8_t *scriptbuf,int32_t scriptlen,int32_t height,uint256 txhash,int32_t i,int32_t j,uint64_t *voutmaskp,int32_t *specialtxp,int32_t *notarizedheightp,uint64_t value,int32_t notarized,uint64_t signedmask,uint32_t timestamp);
-unsigned int lwmaGetNextPOSRequired(const CBlockIndex* pindexLast, const Consensus::Params& params);
-bool EnsureWalletIsAvailable(bool avoidException);
-extern bool fRequestShutdown;
-extern CScript KOMODO_EARLYTXID_SCRIPTPUB;
-
-uint32_t komodo_heightstamp(int32_t height);
 
 //#define issue_curl(cmdstr) bitcoind_RPC(0,(char *)"curl",(char *)"http://127.0.0.1:7776",0,0,(char *)(cmdstr))
