+++ src/komodo_gateway.h
@@ -682,8 +682,6 @@ int32_t komodo_bannedset(int32_t *indallvoutsp,uint256 *array,int32_t max)
     return(i);
 }
 
-void komodo_passport_iteration();
-
 int32_t komodo_check_deposit(int32_t height,const CBlock& block,uint32_t prevtime) // verify above block is valid pax pricing
 {
     static uint256 array[64]; static int32_t numbanned,indallvouts;
@@ -1425,8 +1423,6 @@ int32_t komodo_faststateinit(struct komodo_state *sp,char *fname,char *symbol,ch
     }
     return(-1);
 }
-
-uint64_t komodo_interestsum();
 
 void komodo_passport_iteration()
 {
