+++ src/komodo_globals.h
@@ -103,7 +103,6 @@ uint64_t ASSETCHAINS_COMMISSION,ASSETCHAINS_SUPPLY = 10,ASSETCHAINS_FOUNDERS_REW
 uint32_t KOMODO_INITDONE;
 char KMDUSERPASS[8192+512+1],BTCUSERPASS[8192]; uint16_t KMD_PORT = 7771,BITCOIND_RPCPORT = 7771, DEST_PORT;
 uint64_t PENDING_KOMODO_TX;
-extern int32_t KOMODO_LOADINGBLOCKS;
 unsigned int MAX_BLOCK_SIGOPS = 20000;
 
 int32_t KOMODO_TESTNODE, KOMODO_SNAPSHOT_INTERVAL;
