+++ src/rpc/blockchain.cpp
@@ -46,14 +46,16 @@
 #include "cc/CCinclude.h"
 #include "cc/CCPrices.h"
 
-using namespace std;
-
-extern int32_t KOMODO_INSYNC;
-extern void TxToJSON(const CTransaction& tx, const uint256 hashBlock, UniValue& entry);
-void ScriptPubKeyToJSON(const CScript& scriptPubKey, UniValue& out, bool fIncludeHex);
-int32_t komodo_notarized_height(int32_t *prevMoMheightp, uint256 *hashp, uint256 *txidp);
 #include "komodo_defs.h"
 #include "komodo_structs.h"
+
+using namespace std;
+
+//extern int32_t KOMODO_INSYNC;
+//extern void TxToJSON(const CTransaction& tx, const uint256 hashBlock, UniValue& entry);
+//void ScriptPubKeyToJSON(const CScript& scriptPubKey, UniValue& out, bool fIncludeHex);
+//int32_t komodo_notarized_height(int32_t *prevMoMheightp, uint256 *hashp, uint256 *txidp);
+
 
 double GetDifficultyINTERNAL(const CBlockIndex* blockindex, bool networkDifficulty)
 {
@@ -288,7 +290,7 @@ UniValue blockToDeltasJSON(const CBlock& block, const CBlockIndex* blockindex)
     return result;
 }
 
-UniValue blockToJSON(const CBlock& block, const CBlockIndex* blockindex, bool txDetails = false)
+UniValue blockToJSON(const CBlock& block, const CBlockIndex* blockindex, bool txDetails)
 {
     UniValue result(UniValue::VOBJ);
     uint256 notarized_hash, notarized_desttxid; int32_t prevMoMheight, notarized_height;
@@ -451,7 +453,7 @@ bool mytxid_inmempool(uint256 txid)
     return(false);
 }
 
-UniValue mempoolToJSON(bool fVerbose = false)
+UniValue mempoolToJSON(bool fVerbose)
 {
     if (fVerbose)
     {
